requirements.txt
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
flask-mongoengine==1.0.0
Flask-WTF==0.15.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
pymongo==3.11.4
python-dotenv==0.18.0
Werkzeug==2.0.1
WTForms==2.3.3
gunicorn
